Refactor common S3 functions into dedicated module.
